+++ src/test/java/com/google/sps/data/RetrieveActivityFromDatastoreTest.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps;
+package com.google.sps.data;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
